Created binify_monolithic.m, which separates a monolithic array into smaller bins by their indices.

+++ data/esl_accel/import_file.m
@@ -23,6 +23,6 @@ while true
 end
 
 % why is there no str2int? Did I just not look?
-accel = int16(accel);
+accel = int32(accel);
 fclose(fp);
 end
